feat(DrinkCard):created separate component for the drinksCard
