+++ src/components/Navbar.tsx
@@ -2,7 +2,7 @@ import {Link} from 'react-router-dom';
 
 const Navbar = () => {
     return (
-        <nav className="navbar navbar-expand-lg navbar-dark bg-dark mt-5">
+        <nav className="navbar navbar-expand-lg navbar-dark bg-dark ">
             {/*logo*/}
             <Link to='/' className='navbar-brand'>Drinks</Link>
             {/*menu*/}
