+++ src/pages/Homepage.tsx
@@ -1,10 +1,10 @@
 import {useEffect, useContext} from 'react';
 import {GlobalContext} from '../context/GlobalContext';
-import {Link,useHistory} from 'react-router-dom';
+import DrinkCard from '../components/DrinksCard';
 import '../App.css';
 const Homepage = () => {
     const { getDrinks, drinks} = useContext(GlobalContext)
-    const history = useHistory();
+    
     useEffect(() => {
         getDrinks() ;
      
@@ -23,34 +23,8 @@ const Homepage = () => {
                 return(
                     <div className= 'col-sm-12 col-md-3 mb-3' key={index}>
                         {/***drink card */}
-                        <div className='card '>
-                     <span onClick={() => history.push(`/cocktails/${drink.idDrink}`)}>
-                         {/* image top*/}
-                   
-                     <div className='dr-bg-img' style = {{ backgroundImage: `url(${drink.strDrinkThumb})`}}>
-                  
-                  </div>
-          {/* card body */}
-          <div className='card-body'>
-              <h4 className='card-title text-center'>
-                  <Link to={`/products/${drink.idDrink}`}>{drink.strDrink}</Link></h4>
-              <div className='text-center'>
-              <h5 className='card-title'>{drink.strCategory}</h5>
-              <h6 className='card-title'>{drink.strAlcoholic}</h6>
-              <h6 className='card-title'>{drink.strGlass}</h6>
-               
-             
-
-              </div>
-              
-          </div>
-
-            </span>
-       
-        
-        </div>
-
-        </div>
+                        <DrinkCard drink={drink}/>
+                   </div>
                 )
             })}
